requirements.txt
# Streamlit UI (NLP-to-SQL Test UI)
streamlit==1.47.1

# Core Dependencies
fastapi==0.109.0
uvicorn==0.27.0
flask==2.2.5
gunicorn==22.0.0
python-dotenv>=1.0.0
pydantic>=2.10.6
pydantic-settings==2.2.1
openai==1.30.1
anthropic==0.8.0

# Data Processing
pandas==2.1.1
numpy==1.26.2
duckdb>=0.10.0
pyyaml>=6.0.2

# Dashboard and Visualization (for Decision Studio UI)
dash==2.13.0
plotly==5.17.0

# Testing
pytest>=8.2.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.2
pytest-mock>=3.11.0

# Utilities
python-jose==3.3.0
passlib>=1.7.4
bcrypt==4.0.1
httpx==0.25.2
colorama==0.4.6
tabulate==0.9.0

# Email Integration (for HITL notifications)
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0

# Dev Tooling
pre-commit==3.7.1
